+++ src/components/atoms/index.js
@@ -4,6 +4,7 @@ export { default as Fieldset } from './Fieldset/Fieldset';
 export { default as Input } from './Input/Input';
 export { default as Logo } from './Logo/Logo';
 export { default as Paragraph } from './Paragraph/Paragraph';
+export { default as SectionWrapper } from './SectionWrapper/SectionWrapper';
 export { default as Select } from './Select/Select';
 export { default as SubTitle } from './SubTitle/SubTitle';
 export { default as Title } from './Title/Title';
