+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Root from './views/Root';
+import Root from 'views/Root';
 
 ReactDOM.render(
   <React.StrictMode>
