+++ src/templates/ItemTemplate/ItemTemplate.js
@@ -1,5 +1,4 @@
 ﻿import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
 import { Formik } from 'formik';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
@@ -10,100 +9,25 @@ import {
   Input,
   ButtonIcon,
   Select,
-} from '../components/atoms';
+} from 'components/atoms';
 import {
   removeItem as removeItemAction,
   changeItem as changeItemAction,
   addStock as addStockAction,
   subStock as subStockAction,
-} from '../actions/pantryActions';
-import Header from '../components/molecules/Header/Header';
-import { deleteIcon, applyIcon, plusIcon, minusIcon } from '../assets/icons';
-
-const ItemWrapper = styled.section`
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-  padding: 0 25px;
-
-  @media (max-width: ${({ theme }) => theme.mediaBreaks.mobile}px) {
-    padding: 10px 10px 0;
-    width: 100%;
-  }
-`;
-
-const StyledSettings = styled.form`
-  height: 100%;
-  display: grid;
-  align-items: center;
-  grid-template-columns: 1fr 100px;
-  gap: 15px;
-  align-items: center;
-  font-size: 2.2rem;
-
-  input {
-    font-size: 2rem;
-  }
-
-  .grid-settings-6 {
-    grid-column: 2/3;
-    grid-row: 3/4;
-  }
-  .grid-settings-7 {
-    grid-column: 1/2;
-    grid-row: 4/5;
-    justify-self: end;
-  }
-  .grid-settings-8 {
-    grid-column: 2/3;
-    grid-row: 4/5;
-    justify-self: center;
-  }
-`;
-
-const StyledStock = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 80px 80px;
-  gap: 27px;
-  margin-bottom: 30px;
-
-  label {
-    font-size: 3rem;
-  }
-  input {
-    font-size: 3rem;
-  }
-`;
-
-const StyledDelete = styled.div`
-  margin-top: 30px;
-  margin-bottom: 20px;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-`;
-
-const StyledApplyIcon = styled(ButtonIcon)`
-  background-size: 65%;
-`;
-
-const ChangeStockWrapper = styled.div`
-  display: grid;
-  align-items: center;
-  grid-template-columns: 1fr 1fr;
-`;
-
-const ButtonAdd = styled(ButtonIcon)`
-  border-radius: 15px 0 0 15px;
-  height: 100%;
-  background-size: 65%;
-`;
-const ButtonSub = styled(ButtonIcon)`
-  border-radius: 0 15px 15px 0;
-  height: 100%;
-  background-size: 65%;
-`;
+} from 'actions/pantryActions';
+import { Header } from 'components/molecules';
+import {
+  Wrapper,
+  Stock,
+  ChangeStockWrapper,
+  ButtonAdd,
+  ButtonSub,
+  Settings,
+  ApplyIcon,
+  Delete,
+} from './styles/StyledItemTemplate';
+import { deleteIcon, applyIcon, plusIcon, minusIcon } from '../../assets/icons';
 
 const ItemTemplate = ({
   id,
@@ -126,16 +50,16 @@ const ItemTemplate = ({
   }, [stock]);
 
   return (
-    <ItemWrapper>
+    <Wrapper>
       <Header titleText={name} subTitleText={`category: ${category}`} />
-      <StyledStock>
+      <Stock>
         <label htmlFor="stock">Stock</label>
         <Input id="stock" defaultValue={stockState} settings />
         <ChangeStockWrapper>
           <ButtonAdd icon={plusIcon} onClick={() => addStock(id)} />
           <ButtonSub icon={minusIcon} onClick={() => subStock(id)} />
         </ChangeStockWrapper>
-      </StyledStock>
+      </Stock>
       <Fieldset settings legend="settings">
         <Formik
           initialValues={{ maxStock, minStock, unit }}
@@ -153,7 +77,7 @@ const ItemTemplate = ({
             handleSubmit,
             isSubmitting,
           }) => (
-            <StyledSettings onSubmit={handleSubmit}>
+            <Settings onSubmit={handleSubmit}>
               <label className="grid-settings-1" htmlFor="maxStock">
                 Maximum:
               </label>
@@ -193,17 +117,17 @@ const ItemTemplate = ({
               <Paragraph className="grid-settings-7" size="1.6rem">
                 Apply
               </Paragraph>
-              <StyledApplyIcon
+              <ApplyIcon
                 className="grid-settings-8"
                 type="submit"
                 disabled={isSubmitting}
                 icon={applyIcon}
               />
-            </StyledSettings>
+            </Settings>
           )}
         </Formik>
       </Fieldset>
-      <StyledDelete>
+      <Delete>
         <Paragraph style={{ marginRight: '20px' }} size="1.6rem">
           delete item
         </Paragraph>
@@ -214,8 +138,8 @@ const ItemTemplate = ({
             history.push('/');
           }}
         />
-      </StyledDelete>
-    </ItemWrapper>
+      </Delete>
+    </Wrapper>
   );
 };
