+++ src/templates/MainTemplate/MainTemplate.js
@@ -1,8 +1,8 @@
 ﻿import React from 'react';
 import { ThemeProvider } from 'styled-components';
 import PropTypes from 'prop-types';
-import GlobalStyles from '../theme/GlobalStyles';
-import { theme } from '../theme/mainTheme';
+import GlobalStyles from '../../theme/GlobalStyles';
+import { theme } from '../../theme/mainTheme';
 
 const MainTemplate = ({ children }) => {
   return (
