+++ src/templates/UserTemplate/UserTemplate.js
@@ -2,10 +2,10 @@
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
-import Sidebar from 'components/organisms/Navbars/Sidebar/Sidebar';
-import TopBar from 'components/organisms/Navbars/Topbar/Topbar';
-import { useWindowWidth } from '../hooks';
-import { signOut as signOutAction } from '../actions/authActions';
+import Sidebar from 'components/organisms/Sidebar/Sidebar';
+import TopBar from 'components/organisms/Topbar/Topbar';
+import useWindowWidth from 'hooks/useWindowWidth';
+import { signOut as signOutAction } from '../../actions/authActions';
 
 const UserWrapper = styled.div`
   padding: 10px;
