+++ src/templates/index.js
@@ -1,0 +1,4 @@
+﻿export { default as InfoTemplate } from './InfoTemplate/InfoTemplate';
+export { default as ItemTemplate } from './ItemTemplate/ItemTemplate';
+export { default as MainTemplate } from './MainTemplate/MainTemplate';
+export { default as UserTemplate } from './UserTemplate/UserTemplate';
