+++ src/views/index.js
@@ -1,17 +1,7 @@
-﻿import ItemPage from './ItemPage';
-import NewPage from './NewPage';
-import PantryPage from './PantryPage';
-import SettingsPage from './SettingsPage';
-import ShopListPage from './ShopListPage';
-import SignIn from './SignIn';
-import SignUp from './SignUp';
-
-export {
-  ItemPage,
-  NewPage,
-  PantryPage,
-  SettingsPage,
-  ShopListPage,
-  SignIn,
-  SignUp,
-};
+﻿export { default as ItemPage } from './ItemPage/ItemPage';
+export { default as NewPage } from './NewPage/NewPage';
+export { default as PantryPage } from './PantryPage/PantryPage';
+export { default as SettingsPage } from './SettingsPage/SettingsPage';
+export { default as ShopListPage } from './ShopListPage/ShopListPage';
+export { default as SignIn } from './SignIn/SignIn';
+export { default as SignUp } from './SignUp/SignUp';
